// damn/types.ts
export type Product = {
	id: string;
	name: string;
	slug?: string;
	description: string;

	facetValues: FacetValue[];
	featuredAsset: FeaturedAsset;
	assets: Asset[];
	variants: Variant[];
};




type Facet = {
	id: string;
	code: string;
	name: string;
};

type FacetValue = {
	facet: Facet;
	id: string;
	code: string;
	name: string;
};

type FeaturedAsset = {
	id: string;
	preview: string;
};

type Asset = {
	id: string;
	preview: string;
};

export type Variant = {
	id: string;
	name: string;
	priceWithTax: number;
	currencyCode: CurrencyCode;
	sku: string;
	stockLevel: string;
	featuredAsset?: any;
};